refactor: изменил структуру папки components. Переименовал файлы и переменные
