+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { Translation } from './components/Translation';
+import { Translation } from './components/Translation/Translation';
 
 const queryClient = new QueryClient();
